--- a/.gitignore
+++ b/.gitignore
@@ -6,6 +6,7 @@ coverage.html
 *.heapprofile
 go.work
 go.work.sum
+report.json
 
 # Common in OSs and IDEs
 .idea